+++ dst/app/middlewares/notFoundHandler.js
@@ -4,7 +4,7 @@
  * @module middlewares.notFoundHandler
  */
 Object.defineProperty(exports, "__esModule", { value: true });
-const waiter_domain_1 = require("@motionpicture/waiter-domain");
+const domain_1 = require("@waiter/domain");
 exports.default = (req, __, next) => {
-    next(new waiter_domain_1.factory.errors.NotFound(`router for [${req.originalUrl}]`));
+    next(new domain_1.factory.errors.NotFound(`router for [${req.originalUrl}]`));
 };